Bugfix
Change stop function. It did set the tourKey back to '_default' but this
resulted in the 'stop' event not being emitted for the current tour.

Now the stop function leaves the current tourKey value unchanged. It
will change whenever another tour calls the start function or it's
useTourController hook fires.