Multiple tours (#80)
* Fix Step measures when there's a borderRadius set (#42)

* bugfix

Fix bug where pressOnDismiss was not working, it was not being
passed down to SvgMask

* Allow Multiple Tours

Change TourGuideProvider so most of the state / refs are objects that
are keyed by strings (the tour names). Change most functions so they
apply only to one specific key of that object allowing multiple tours
within the same app.

Update the useTourGuideController hook to take in an optional string key
to designate the tour.  The hook returns methods that are keyed to that
specific tour as well as components that are keyed to that tour.

If no tourKey is supplied to the hook (existing implementations) then
the default key is used.

Update README.md with examples

* Bugfix

Change stop function. It did set the tourKey back to '_default' but this
resulted in the 'stop' event not being emitted for the current tour.

Now the stop function leaves the current tourKey value unchanged. It
will change whenever another tour calls the start function or it's
useTourController hook fires.

Co-authored-by: Emilio Rodriguez <emiliorodriguez@gmail.com>
Co-authored-by: Ben Steel <ben@think-now.com>
Co-authored-by: Xavier Carpentier <xcapetir@gmail.com>